+++ ApplicationExamples/CompressibleNavierStokes/AMR/VarianceHelper.h
@@ -4,6 +4,6 @@
 #include <utility>
 
 std::pair<double, double> mergeVariance(double mean0, double mean1, double var0,
-                                        double var1, int count0, int count1);
+                                        double var1, double count0, double count1);
 
 #endif  // COMPRESSIBLENAVIERSTOKES_VARIANCEHELPER_H
